+++ extensions/esbuild-extension-common.mts
