+++ extensions/esbuild-webview-common.mts
@@ -2,27 +2,22 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-// @ts-check
+
 /**
- * @fileoverview Common build script for extension scripts used in in webviews.
+ * Common build script for extension scripts used in in webviews.
  */
 import path from 'node:path';
 import esbuild from 'esbuild';
 
-/**
- * @typedef {Partial<import('esbuild').BuildOptions> & {
- * 	entryPoints: string[] | Record<string, string> | { in: string, out: string }[];
- * 	outdir: string;
- * }} BuildOptions
- */
+export type BuildOptions = Partial<esbuild.BuildOptions> & {
+	entryPoints: string[] | Record<string, string> | { in: string; out: string }[];
+	outdir: string;
+};
 
 /**
  * Build the source code once using esbuild.
- *
- * @param {BuildOptions} options
- * @param {(outDir: string) => unknown} [didBuild]
  */
-async function build(options, didBuild) {
+async function build(options: BuildOptions, didBuild?: (outDir: string) => unknown): Promise<void> {
 	await esbuild.build({
 		bundle: true,
 		minify: true,
@@ -38,11 +33,8 @@ async function build(options, didBuild) {
 
 /**
  * Build the source code once using esbuild, logging errors instead of throwing.
- *
- * @param {BuildOptions} options
- * @param {(outDir: string) => unknown} [didBuild]
  */
-async function tryBuild(options, didBuild) {
+async function tryBuild(options: BuildOptions, didBuild?: (outDir: string) => unknown): Promise<void> {
 	try {
 		await build(options, didBuild);
 	} catch (err) {
@@ -50,17 +42,16 @@ async function tryBuild(options, didBuild) {
 	}
 }
 
-/**
- * @param {{
- * 	srcDir: string;
- *  outdir: string;
- *  entryPoints: string[] | Record<string, string> | { in: string, out: string }[];
- * 	additionalOptions?: Partial<import('esbuild').BuildOptions>
- * }} config
- * @param {string[]} args
- * @param {(outDir: string) => unknown} [didBuild]
- */
-export async function run(config, args, didBuild) {
+export async function run(
+	config: {
+		srcDir: string;
+		outdir: string;
+		entryPoints: BuildOptions['entryPoints'];
+		additionalOptions?: Partial<esbuild.BuildOptions>;
+	},
+	args: string[],
+	didBuild?: (outDir: string) => unknown
+): Promise<void> {
 	let outdir = config.outdir;
 	const outputRootIndex = args.indexOf('--outputRoot');
 	if (outputRootIndex >= 0) {
@@ -69,8 +60,7 @@ export async function run(config, args, didBuild) {
 		outdir = path.join(outputRoot, outputDirName);
 	}
 
-	/** @type {BuildOptions} */
-	const resolvedOptions = {
+	const resolvedOptions: BuildOptions = {
 		entryPoints: config.entryPoints,
 		outdir,
 		logOverride: {
