+++ extensions/ipynb/esbuild.notebook.mts
@@ -2,16 +2,16 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import * as path from 'node:path';
-import { run } from '../esbuild-extension-common.ts';
+import path from 'node:path';
+import { run } from '../esbuild-webview-common.mts';
 
-const srcDir = path.join(import.meta.dirname, 'src');
-const outDir = path.join(import.meta.dirname, 'dist');
+const srcDir = path.join(import.meta.dirname, 'notebook-src');
+const outDir = path.join(import.meta.dirname, 'notebook-out');
 
 run({
-	entryPoints: {
-		'extension': path.join(srcDir, 'extension.ts'),
-	},
+	entryPoints: [
+		path.join(srcDir, 'cellAttachmentRenderer.ts'),
+	],
 	srcDir,
 	outdir: outDir,
 }, process.argv);
