+++ extensions/markdown-language-features/esbuild.browser.mts
@@ -4,24 +4,37 @@
  *--------------------------------------------------------------------------------------------*/
 import * as fs from 'node:fs';
 import * as path from 'node:path';
-import { run } from '../esbuild-extension-common.ts';
+import { run } from '../esbuild-extension-common.mts';
 
 const srcDir = path.join(import.meta.dirname, 'src');
-const outDir = path.join(import.meta.dirname, 'dist');
+const outDir = path.join(import.meta.dirname, 'dist', 'browser');
 
 /**
  * Copy the language server worker main file to the output directory.
  */
 async function copyServerWorkerMain(outDir: string): Promise<void> {
-	const srcPath = path.join(import.meta.dirname, 'node_modules', 'vscode-markdown-languageserver', 'dist', 'node', 'workerMain.js');
+	const srcPath = path.join(import.meta.dirname, 'node_modules', 'vscode-markdown-languageserver', 'dist', 'browser', 'workerMain.js');
 	const destPath = path.join(outDir, 'serverWorkerMain.js');
 	await fs.promises.copyFile(srcPath, destPath);
 }
 
 run({
 	entryPoints: {
-		'extension': path.join(srcDir, 'extension.ts'),
+		'extension': path.join(srcDir, 'extension.browser.ts'),
 	},
 	srcDir,
 	outdir: outDir,
+	additionalOptions: {
+		platform: 'browser',
+		format: 'cjs',
+		alias: {
+			'path': 'path-browserify',
+		},
+		define: {
+			'process.platform': JSON.stringify('web'),
+			'process.env': JSON.stringify({}),
+			'process.env.BROWSER_ENV': JSON.stringify('true'),
+		},
+		tsconfig: path.join(import.meta.dirname, 'tsconfig.browser.json'),
+	},
 }, process.argv, copyServerWorkerMain);
