+++ extensions/markdown-language-features/esbuild.notebook.mts
@@ -1,0 +1,17 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+import path from 'path';
+import { run } from '../esbuild-webview-common.mts';
+
+const srcDir = path.join(import.meta.dirname, 'notebook');
+const outDir = path.join(import.meta.dirname, 'notebook-out');
+
+run({
+	entryPoints: [
+		path.join(srcDir, 'index.ts'),
+	],
+	srcDir,
+	outdir: outDir,
+}, process.argv);
