+++ extensions/markdown-language-features/esbuild.webview.mts
@@ -2,16 +2,16 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-// @ts-check
 import path from 'path';
-import { run } from '../esbuild-webview-common.mjs';
+import { run } from '../esbuild-webview-common.mts';
 
-const srcDir = path.join(import.meta.dirname, 'notebook');
-const outDir = path.join(import.meta.dirname, 'notebook-out');
+const srcDir = path.join(import.meta.dirname, 'preview-src');
+const outDir = path.join(import.meta.dirname, 'media');
 
 run({
 	entryPoints: [
 		path.join(srcDir, 'index.ts'),
+		path.join(srcDir, 'pre'),
 	],
 	srcDir,
 	outdir: outDir,
