+++ extensions/markdown-math/package.json
@@ -108,7 +108,7 @@
   "scripts": {
     "compile": "npm run build-notebook",
     "watch": "npm run build-notebook",
-    "build-notebook": "node ./esbuild.mjs"
+    "build-notebook": "node ./esbuild.mts"
   },
   "devDependencies": {
     "@types/markdown-it": "^0.0.0",
