+++ extensions/markdown-math/tsconfig.browser.json
@@ -1,0 +1,9 @@
+{
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+		"types": [],
+		"typeRoots": [
+			"./node_modules/@types"
+		]
+	}
+}
