+++ extensions/mermaid-chat-features/esbuild.webview.mts
@@ -2,16 +2,16 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-// @ts-check
 import path from 'path';
-import { run } from '../esbuild-webview-common.mjs';
+import { run } from '../esbuild-webview-common.mts';
 
-const srcDir = path.join(import.meta.dirname, 'preview-src');
-const outDir = path.join(import.meta.dirname, 'media');
+const srcDir = path.join(import.meta.dirname, 'chat-webview-src');
+const outDir = path.join(import.meta.dirname, 'chat-webview-out');
 
 run({
 	entryPoints: {
 		'index': path.join(srcDir, 'index.ts'),
+		'index-editor': path.join(srcDir, 'index-editor.ts'),
 		'codicon': path.join(import.meta.dirname, 'node_modules', '@vscode', 'codicons', 'dist', 'codicon.css'),
 	},
 	srcDir,
