+++ extensions/simple-browser/package.json
@@ -71,7 +71,7 @@
     "watch": "npm run build-preview && gulp watch-extension:simple-browser",
     "vscode:prepublish": "npm run build-ext && npm run build-preview",
     "build-ext": "node ../../node_modules/gulp/bin/gulp.js --gulpfile ../../build/gulpfile.extensions.mjs compile-extension:simple-browser ./tsconfig.json",
-    "build-preview": "node ./esbuild-preview.mjs",
+    "build-preview": "node ./esbuild-preview.mts",
     "compile-web": "npx webpack-cli --config extension-browser.webpack.config --mode none",
     "watch-web": "npx webpack-cli --config extension-browser.webpack.config --mode none --watch --info-verbosity verbose"
   },
